+++ packages/morpheme_http/lib/src/cache_strategy/model/cache_wrapper.dart
@@ -1,54 +1,56 @@
 import 'dart:convert';
 
 import 'package:http/http.dart';
+import 'package:objectbox/objectbox.dart';
 
+@Entity()
 final class CacheWrapper {
   CacheWrapper({
+    required this.key,
     required this.cacheDate,
-    required this.response,
+    this.response,
   });
 
+  @Id()
+  int id = 0;
+
+  @Index(type: IndexType.value)
+  @Unique(onConflict: ConflictStrategy.replace)
+  final String key;
+
   final int cacheDate;
-  final Response response;
 
-  Map<String, dynamic> toMap() {
-    return {
-      'cacheDate': cacheDate,
-      'response': {
-        'body': response.body,
-        'status_code': response.statusCode,
-        'headers': response.headers,
-        'is_redirect': response.isRedirect,
-        'persistent_connection': response.persistentConnection,
-        'reason_phrase': response.reasonPhrase,
-        'request': {
-          'url': response.request?.url.toString(),
-          'method': response.request?.method,
+  @Transient()
+  Response? response;
+
+  String get dbResponse => json.encode(
+        {
+          'body': response?.body,
+          'status_code': response?.statusCode,
+          'headers': response?.headers,
+          'is_redirect': response?.isRedirect,
+          'persistent_connection': response?.persistentConnection,
+          'reason_phrase': response?.reasonPhrase,
+          'request': {
+            'url': response?.request?.url.toString(),
+            'method': response?.request?.method,
+          },
         },
-      },
-    };
-  }
+      );
 
-  factory CacheWrapper.fromMap(Map<String, dynamic> map) {
-    return CacheWrapper(
-      cacheDate: map['cacheDate']?.toInt() ?? 0,
-      response: Response(
-        map['response']['body'] ?? '',
-        map['response']['status_code'] ?? 400,
-        headers: Map<String, String>.from(map['response']['headers'] ?? {}),
-        isRedirect: map['response']['is_redirect'],
-        persistentConnection: map['response']['persistent_connection'],
-        reasonPhrase: map['response']['reason_phrase'],
-        request: Request(
-          map['response']['request']['method'] ?? '',
-          Uri.parse(map['response']['request']['url'] ?? ''),
-        ),
+  set dbResponse(String value) {
+    final map = json.decode(value);
+    response = Response(
+      map['body'] ?? '',
+      map['status_code'] ?? 400,
+      headers: Map<String, String>.from(map['headers'] ?? {}),
+      isRedirect: map['is_redirect'],
+      persistentConnection: map['persistent_connection'],
+      reasonPhrase: map['reason_phrase'],
+      request: Request(
+        map['request']['method'] ?? '',
+        Uri.parse(map['request']['url'] ?? ''),
       ),
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory CacheWrapper.fromJson(String source) =>
-      CacheWrapper.fromMap(json.decode(source));
 }