+++ packages/morpheme_http/lib/src/cache_strategy/storage/storage.dart
@@ -1,16 +1,18 @@
+import 'package:morpheme_http/src/cache_strategy/cache_strategy.dart';
+
 abstract interface class Storage {
   /// A function that takes a key and a value and returns a future.
-  Future<void> write(String key, String value);
+  Future<int?> write(CacheWrapper value);
 
   /// Returning a future that will return a string or null.
-  Future<String?> read(String key);
+  Future<CacheWrapper?> read(String key);
 
   /// Deleting a key from the storage.
-  Future<void> delete(String key);
+  Future<int?> delete(String key);
 
   /// Counting the number of keys in the storage.
   Future<int> count({String? prefix});
 
   /// A function that takes a string and returns a future.
-  Future<void> clear({String? prefix});
+  Future<int?> clear({String? prefix});
 }