+++ packages/morpheme_responsive/lib/morpheme_responsive.dart
@@ -1,5 +1,3 @@
-library morpheme_responsive;
-
 export 'src/extensions/morpheme_responsive_extension.dart';
 export 'src/model/morpheme_breakpoint.dart';
 export 'src/model/morpheme_inherited_breakpoint.dart';